requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
openai==1.30.1
pdfplumber==0.11.0
python-multipart==0.0.9
python-dotenv==1.0.1
motor==3.4.0
pymongo==4.7.2
dnspython==2.6.1
